created root screen and moved theming up to main activity
